Add correct test file,, revert unrelated change
